fix(lint): clippy + workspace integration test fixes for u16 migration
- `less_than::execution`: generalize `InterpreterExecutor` /
  `InterpreterMeteredExecutor` impls over `NUM_LIMBS` so the u16-shaped
  `LessThanExecutor<_, 4, 16>` instantiation satisfies the bounds used
  by `air_test_with_min_segments` in `openvm-deferral-integration-tests`
  and `openvm-continuations`. The body of these impls doesn't depend on
  NUM_LIMBS (it dispatches to a host-side handler that operates on u64
  values reconstructed from the bus), so the relaxation is mechanical.
- `is_eq.rs:750`: `i % 2 == 0` → `i.is_multiple_of(2)` (clippy 1.91 lint).
- `eq_mod_u16.rs:57,166`, `vec_heap_branch.rs:159`, `vec_heap_u16.rs:212`:
  drop unnecessary `as usize` casts where the source is already `usize`.
- `eq_mod_u16.rs:166`: use `+=` instead of `acc = acc + ...`.
- `jal_lui/tests.rs:124,131,565`: drop unnecessary `i32` casts; annotate
  the bind site type instead.
- `sha2/tests.rs:21`: gate the `VariableRangeCheckerChip` import behind
  `cfg(feature = "cuda")` since it's only used by the GPU harness.

`cargo clippy --workspace --all-targets --tests -- -D warnings` and
`cargo check --workspace --all-targets --tests` both clean.
`cargo +nightly fmt --all -- --check` clean.

+++ extensions/riscv-adapters/src/vec_heap_branch.rs
@@ -156,7 +156,7 @@ impl<
         let e = AB::F::from_u32(RV64_MEMORY_AS);
         // Reads from heap. After Pattern B, `READ_SIZE` is u16-cell count (= BLOCK_FE_WIDTH=4
         // for an 8-byte chunk); each block's bus pointer advances by 2*READ_SIZE bytes.
-        const BUS_PTR_SCALE: usize = openvm_circuit::arch::BUS_PTR_SCALE as usize;
+        const BUS_PTR_SCALE: usize = openvm_circuit::arch::BUS_PTR_SCALE;
         let bytes_per_block = READ_SIZE * BUS_PTR_SCALE;
         for (address, reads, reads_aux) in izip!(rs_val_f, ctx.reads, &cols.reads_aux) {
             for (i, (read, aux)) in zip(reads, reads_aux).enumerate() {
